// _fruitables_notification/src/common/enum/common.enum.ts
export enum STATUS {
  PENDING = 'Pending',
  INPROCESS = 'Inprocess',
  EXECUTED = 'Executed',
  FAILED = 'Failed',
}

export enum TYPE {
  API = 'API',
  ADMIN = 'Admin',
  FRONT = 'Front',
  NOTIFICATIONS = 'Notifications',
}

export enum MODE {
  LIVE = 'live',
  SANDBOX = 'sandbox',
}
export enum CODE {
  USER = 'USER',
  ALERT = 'ALERT',
}

export enum SILENT {
  YES = 'Yes',
  NO = 'No',
}

export enum DEVICE {
  IOS = 'iOS',
  ANDROID = 'Android',
}